* AmazonのURLを強制的に書き換えるようにした * デバッグ出力にムダにOverflowExceptionのメッセージが出ていたのでlongにするようにした
git-svn-id: http://svn.misuzilla.org/etcetera/TwitterIrcGateway/branches/hosting@654 07c13e24-2e35-da11-a9e7-00b0d02dc07a

+++ TwitterIrcGatewayCore/AddIns/DLRIntegration/Samples/scraping.py
@@ -124,7 +124,7 @@ class Scraping(Object):
 			# Status
 			s.Source    = self.re_source.search(status).group(1)
 			s.Text      = Utility.UnescapeCharReference(self.re_tag.sub(r"", self.re_anchor.sub(r"\1", self.re_content.search(status).group(1))))
-			s.Id        = int(self.re_status_id.search(status).group(1), 10)
+			s.Id        = long(self.re_status_id.search(status).group(1), 10)
 			s.CreatedAt = DateTime.Now
 			
 			#Trace.WriteLine(s.ToString())
